Moved Footer helpers to helpers-file
