+++ src/components/footer/Footer.js
@@ -1,12 +1,12 @@
 import React from "react";
 import "./Footer.css";
 import "../../App.css";
-import SocialIcon from "./SocialIcon";
+import SocialIcon from "./helpers/SocialIcon";
 import {ReactComponent as FacebookIcon} from "../../assets/icons/facebook.svg";
 import {ReactComponent as InstagramIcon} from "../../assets/icons/instagram.svg";
 import {ReactComponent as LinkedInIcon} from "../../assets/icons/linkedin.svg";
 import {ReactComponent as TwitterIcon} from "../../assets/icons/twitter.svg";
-import FooterListItem from "./FooterListItem";
+import FooterListItem from "./helpers/FooterListItem";
 
 
 function Footer() {
