+++ src/components/footer/helpers/SocialIcon.js
@@ -1,5 +1,5 @@
 import React from "react";
-import "./Footer.css";
+import "../Footer.css";
 import {Link} from "react-router-dom";
